2014/05/20(1)

+++ Workspace/rebok2j/src/main/java/org/rebok2j/dao/AutorizationDao.java
@@ -10,6 +10,6 @@ public interface AutorizationDao {
 
     void remove(Autorization autorization);
 
-    Autorization findById(Integer id);
+    Autorization findById(Long id);
 
 }
