+++ Workspace/rebok2j/src/main/java/org/rebok2j/dao/CourierCompanyDao.java
@@ -10,6 +10,6 @@ public interface CourierCompanyDao {
 
     void remove(CourierCompany courierCompany);
 
-    CourierCompany findById(Integer id);
+    CourierCompany findById(Long id);
 
 }
